JENKINS-46132 - Fix the test for 2.88.1 baseline.
The underlying implementation of some of the the FilePath functionality
have changed to NIO. Therefore a `NoSuchFileException` is thrown instead
of the `FileNotFoundException` when file is missing. `FilePathContent`
is update to handle both in the same way. `FileContent` has not been
update since it is still using the old File IO.