Upgrade parent pom & fix InputStream leak & fix tests
The newer jenkins-test-harness provided with this change does not
any longer swallow exceptions on test teardown. See
https://github.com/jenkinsci/jenkins-test-harness/pull/166

This reveals problems in the implementation and tests. Under Windows
tests failed because some folders could not be removed due to open
file handles.

Fixes:

1) In `BaseFileContent` the `InputStream` was not closed correctly
in every case.

2) The `SemaphoreStep` was not used properly and anyway not required
in most cases. It has to be used correctly which requires 1-2 lines
of code in every test to not leave the environment in a bad state.
The platform-independent `echo` step is sufficient and requires no
extra code. Using no step at all may also be an option.