process the NICs in parallel to reduce the wallclock time
On windows it is normal to have a lot of NICs (for example my laptop enumerates 58 NICs)

it takes 900ms to build the string for this which is outside the time that the callable is allowed to take.

Whilst on Unix like systems you will likely have a lot fewer NICs the resulting code may take a bit longer
 and it relies on the shared fork-join threadpool this is probably still worthwhile.

On my System the callable now takes ~390ms which is half the time it did before and under the 500ms that timesout the callable.