Improve output on NodeRemoteDirectoryComponentTest failures
The NodeRemoteDirectoryComponentTest failed 3 of 10 runs on my Linux
computer with Java 21 and the command:

  mvn -DforkCount=1C test

I see a similar failure rate on my CI server.

I don't understand the cause of the failures. This pull request improves
the failure message with more context when the failure occurs.

Now the failure messages report the contents of the collection that is
failing the assertion:

[ERROR] Failures:
[ERROR]   NodeRemoteDirectoryComponentTest.addContents:37
Expected: a collection containing a string matching the pattern 'nodes/slave/slave0/remote/support/.*.log'
     but: was empty

[ERROR]   NodeRemoteDirectoryComponentTest.addContentsWithIncludes:71
Expected: a collection containing a string matching the pattern 'nodes/slave/agent1/remote/support/.*.log'
     but: was empty